+++ lib/gretel.rb
@@ -1,12 +1,16 @@
 require 'gretel/version'
+require 'gretel/resettable'
 require 'gretel/crumbs'
 require 'gretel/crumb'
 require 'gretel/link'
+require 'gretel/renderer'
 require 'gretel/view_helpers'
 require 'gretel/deprecated'
 
 module Gretel
   class << self
+    include Resettable
+
     # Returns the path from with breadcrumbs are loaded. Default is +config/breadcrumbs.rb+.
     def breadcrumb_paths
       @breadcrumb_paths ||= [Rails.root.join("config", "breadcrumbs.rb"), Rails.root.join("config", "breadcrumbs", "**", "*.rb")]
@@ -30,6 +34,7 @@ module Gretel
     # Shows a deprecation warning.
     def show_deprecation_warning(message)
       return if suppress_deprecation_warnings?
+      message = "[Gretel] #{message}"
       puts message
       Rails.logger.warn message
     end
@@ -39,13 +44,23 @@ module Gretel
       @reload_environments ||= ["development"]
     end
 
+    # Registers a style for later use.
+    # 
+    #   Gretel.register_style :ul, { container_tag: :ul, fragment_tag: :li }
+    def register_style(style, options)
+      Gretel::Renderer.register_style style, options
+    end
+
     # Sets the Rails environment names with automatic configuration reload. Default is +["development"]+.
     attr_writer :reload_environments
 
-    # Resets all changes made to +Gretel+ and +Gretel::Crumbs+. Used for testing.
-    def reset!
-      instance_variables.each { |var| remove_instance_variable var }
-      Crumbs.reset!
+    # Yields this +Gretel+ to be configured.
+    # 
+    #   Gretel.configure do |config|
+    #     config.reload_environments << "staging"
+    #   end
+    def configure
+      yield self
     end
   end
 end