+++ lib/gretel/link.rb
@@ -3,7 +3,16 @@ module Gretel
     attr_accessor :key, :text, :url
 
     def initialize(key, text, url)
-      @key, @text, @url = key, text, url
+      # Use accessors so plugins can override their behavior
+      self.key, self.text, self.url = key, text, url
+    end
+
+    def current!
+      @current = true
+    end
+
+    def current?
+      !!@current
     end
   end
 end