+++ test/helper_methods_test.rb
@@ -9,52 +9,54 @@ class HelperMethodsTest < ActionView::TestCase
     Gretel.reset!
   end
 
+  # Breadcrumb generation
+
   test "shows basic breadcrumb" do
     breadcrumb :basic
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <span class="current">About</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">About</span></div>},
                  breadcrumbs
   end
 
   test "shows breadcrumb with root" do
     breadcrumb :with_root
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <span class="current">About</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">About</span></div>},
                  breadcrumbs
   end
 
   test "shows breadcrumb with parent" do
     breadcrumb :with_parent
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <a href="/about">About</a> &gt; <span class="current">Contact</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <a href="/about">About</a> &rsaquo; <span class="current">Contact</span></div>},
                  breadcrumbs
   end
 
   test "shows breadcrumb with autopath" do
     breadcrumb :with_autopath, projects(:one)
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <span class="current">Test Project</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">Test Project</span></div>},
                  breadcrumbs
   end
 
   test "shows breadcrumb with parent object" do
     breadcrumb :with_parent_object, issues(:one)
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <a href="/projects/1">Test Project</a> &gt; <span class="current">Test Issue</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <a href="/projects/1">Test Project</a> &rsaquo; <span class="current">Test Issue</span></div>},
                  breadcrumbs
   end
 
   test "shows multiple links" do
     breadcrumb :multiple_links
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <a href="/about/contact">Contact</a> &gt; <span class="current">Contact form</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <a href="/about/contact">Contact</a> &rsaquo; <span class="current">Contact form</span></div>},
                  breadcrumbs
   end
 
   test "shows multiple links with parent" do
     breadcrumb :multiple_links_with_parent
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <a href="/about">About</a> &gt; <a href="/about/contact">Contact</a> &gt; <span class="current">Contact form</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <a href="/about">About</a> &rsaquo; <a href="/about/contact">Contact</a> &rsaquo; <span class="current">Contact form</span></div>},
                  breadcrumbs
   end
 
   test "shows semantic breadcrumb" do
     breadcrumb :with_root
-    assert_equal %{<div class="breadcrumbs"><div itemscope="" itemtype="http://data-vocabulary.org/Breadcrumb"><a href="/" itemprop="url"><span itemprop="title">Home</span></a></div> &gt; <div itemscope="" itemtype="http://data-vocabulary.org/Breadcrumb"><span class="current" itemprop="title">About</span></div></div>},
-                 breadcrumbs(:semantic => true)
+    assert_equal %{<div class="breadcrumbs"><div itemscope="" itemtype="http://data-vocabulary.org/Breadcrumb"><a href="/" itemprop="url"><span itemprop="title">Home</span></a></div> &rsaquo; <div itemscope="" itemtype="http://data-vocabulary.org/Breadcrumb"><span class="current" itemprop="title">About</span></div></div>},
+                 breadcrumbs(semantic: true)
   end
 
   test "doesn't show root alone" do
@@ -62,10 +64,16 @@ class HelperMethodsTest < ActionView::TestCase
     assert_equal "", breadcrumbs
   end
 
-  test "shows root alone" do
+  test "displays single fragment" do
     breadcrumb :root
     assert_equal %{<div class="breadcrumbs"><span class="current">Home</span></div>},
-                 breadcrumbs(:show_root_alone => true)
+                 breadcrumbs(display_single_fragment: true)
+  end
+
+  test "displays single non-root fragment" do
+    breadcrumb :basic
+    assert_equal %{<div class="breadcrumbs"><span class="current">About</span></div>},
+                 breadcrumbs(autoroot: false, display_single_fragment: true)
   end
 
   test "shows no breadcrumb" do
@@ -74,68 +82,61 @@ class HelperMethodsTest < ActionView::TestCase
 
   test "links current breadcrumb" do
     breadcrumb :with_root
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <a href="/about" class="current">About</a></div>},
-                 breadcrumbs(:link_current => true)
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <a href="/about" class="current">About</a></div>},
+                 breadcrumbs(link_current: true)
   end
 
   test "shows pretext" do
     breadcrumb :basic
-    assert_equal %{<div class="breadcrumbs">You are here: <a href="/">Home</a> &gt; <span class="current">About</span></div>},
-                 breadcrumbs(:pretext => "You are here: ")
+    assert_equal %{<div class="breadcrumbs">You are here: <a href="/">Home</a> &rsaquo; <span class="current">About</span></div>},
+                 breadcrumbs(pretext: "You are here: ")
   end
 
   test "shows posttext" do
     breadcrumb :basic
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <span class="current">About</span> - text after breadcrumbs</div>},
-                 breadcrumbs(:posttext => " - text after breadcrumbs")
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">About</span> - text after breadcrumbs</div>},
+                 breadcrumbs(posttext: " - text after breadcrumbs")
   end
 
   test "autoroot disabled" do
     breadcrumb :basic
-    assert_equal %{<div class="breadcrumbs"><span class="current">About</span></div>},
-                 breadcrumbs(:autoroot => false)
+    assert_equal "", breadcrumbs(autoroot: false)
   end
 
   test "shows separator" do
     breadcrumb :with_root
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">About</span></div>},
-                 breadcrumbs(:separator => " &rsaquo; ")
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &raquo; <span class="current">About</span></div>},
+                 breadcrumbs(separator: " &raquo; ")
   end
 
   test "shows element id" do
     breadcrumb :basic
-    assert_equal %{<div class="breadcrumbs" id="custom_id"><a href="/">Home</a> &gt; <span class="current">About</span></div>},
-                 breadcrumbs(:id => "custom_id")
+    assert_equal %{<div class="breadcrumbs" id="custom_id"><a href="/">Home</a> &rsaquo; <span class="current">About</span></div>},
+                 breadcrumbs(id: "custom_id")
   end
 
   test "shows custom container class" do
     breadcrumb :basic
-    assert_equal %{<div class="custom_class"><a href="/">Home</a> &gt; <span class="current">About</span></div>},
-                 breadcrumbs(:class => "custom_class")
+    assert_equal %{<div class="custom_class"><a href="/">Home</a> &rsaquo; <span class="current">About</span></div>},
+                 breadcrumbs(class: "custom_class")
   end
 
   test "shows custom current class" do
     breadcrumb :basic
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <span class="custom_current_class">About</span></div>},
-                 breadcrumbs(:current_class => "custom_current_class")
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="custom_current_class">About</span></div>},
+                 breadcrumbs(current_class: "custom_current_class")
   end
 
   test "unsafe html" do
     breadcrumb :with_unsafe_html
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <span class="current">Test &lt;strong&gt;bold text&lt;/strong&gt;</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">Test &lt;strong&gt;bold text&lt;/strong&gt;</span></div>},
                  breadcrumbs
   end
 
   test "safe html" do
     breadcrumb :with_safe_html
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <span class="current">Test <strong>bold text</strong></span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">Test <strong>bold text</strong></span></div>},
                  breadcrumbs
-  end
-
-  test "works with legacy breadcrumb rendering method" do
-    breadcrumb :basic
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <span class="current">About</span></div>},
-                 breadcrumb
   end
 
   test "yields a block containing breadcrumb links array" do
@@ -151,31 +152,174 @@ class HelperMethodsTest < ActionView::TestCase
                   [:multiple_links_with_parent, "Contact form", "/about/contact/form"]], out
   end
 
+  test "parent breadcrumb" do
+    breadcrumb :multiple_links_with_parent
+
+    parent = parent_breadcrumb
+    assert_equal [:multiple_links_with_parent, "Contact", "/about/contact"],
+                 [parent.key, parent.text, parent.url]
+  end
+
+  test "yields parent breadcrumb" do
+    breadcrumb :multiple_links_with_parent
+
+    out = parent_breadcrumb do |parent|
+      [parent.key, parent.text, parent.url]
+    end
+    assert_equal [:multiple_links_with_parent, "Contact", "/about/contact"],
+                 out
+  end
+
+  test "parent breadcrumb returns nil if not present" do
+    breadcrumb :basic
+
+    assert_nil parent_breadcrumb(autoroot: false)
+  end
+
+  test "parent breadcrumb yields only if present" do
+    breadcrumb :basic
+
+    out = parent_breadcrumb(autoroot: false) do
+      "yielded"
+    end
+
+    assert_nil out
+  end
+
+  test "sets current on last link in array" do
+    breadcrumb :multiple_links_with_parent
+
+    out = breadcrumbs do |links|
+      links.map(&:current?)
+    end
+
+    assert_equal [false, false, false, true], out
+  end
+
   test "without link" do
     breadcrumb :without_link
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; Also without link &gt; <span class="current">Without link</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; Also without link &rsaquo; <span class="current">Without link</span></div>},
                  breadcrumbs
   end
 
   test "view context" do
     breadcrumb :using_view_helper
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <span class="current">TestTest</span></div>},
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">TestTest</span></div>},
                  breadcrumbs
   end
 
   test "multiple arguments" do
     breadcrumb :with_multiple_arguments, "One", "Two", "Three"
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <a href="/about">First OneOne then TwoTwo then ThreeThree</a> &gt; <span class="current">One and Two and Three</span></div>},
-                 breadcrumb
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <a href="/about">First OneOne then TwoTwo then ThreeThree</a> &rsaquo; <span class="current">One and Two and Three</span></div>},
+                 breadcrumbs
+  end
+
+  test "with_breadcrumb" do
+    breadcrumb :basic
+    
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">About</span></div>},
+                 breadcrumbs
+
+    with_breadcrumb(:with_parent_object, issues(:one)) do
+      assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <a href="/projects/1">Test Project</a> &rsaquo; <span class="current">Test Issue</span></div>},
+                   breadcrumbs
+    end    
+    
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">About</span></div>},
+                 breadcrumbs
   end
 
   test "calling breadcrumbs helper twice" do
     breadcrumb :with_parent
     2.times do
-      assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &gt; <a href="/about">About</a> &gt; <span class="current">Contact</span></div>},
+      assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <a href="/about">About</a> &rsaquo; <span class="current">Contact</span></div>},
                    breadcrumbs
     end
   end
+
+  test "breadcrumb not found" do
+    assert_raises ArgumentError do
+      breadcrumb :nonexistent
+      breadcrumbs
+    end
+  end
+
+  test "current link url is set to fullpath" do
+    self.request = OpenStruct.new(fullpath: "/testpath?a=1&b=2")
+
+    breadcrumb :basic
+    assert_equal "/testpath?a=1&b=2", breadcrumbs { |links| links.last.url }
+  end
+
+  test "calling the breadcrumb method with wrong arguments" do
+    assert_nothing_raised do
+      breadcrumb :basic, test: 1
+    end
+
+    assert_raises ArgumentError do
+      breadcrumb
+    end
+
+    assert_raises ArgumentError do
+      breadcrumb(pretext: "bla")
+    end
+  end
+
+  # Styles
+
+  test "default style" do
+    breadcrumb :basic
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home</a> &rsaquo; <span class="current">About</span></div>},
+                 breadcrumbs
+  end
+
+  test "ordered list style" do
+    breadcrumb :basic
+    assert_equal %{<ol class="breadcrumbs"><li><a href="/">Home</a></li><li class="current">About</li></ol>},
+                 breadcrumbs(style: :ol)
+  end
+
+  test "unordered list style" do
+    breadcrumb :basic
+    assert_equal %{<ul class="breadcrumbs"><li><a href="/">Home</a></li><li class="current">About</li></ul>},
+                 breadcrumbs(style: :ul)
+  end
+
+  test "bootstrap style" do
+    breadcrumb :basic
+    assert_equal %{<ol class="breadcrumb"><li><a href="/">Home</a></li><li class="active">About</li></ol>},
+                 breadcrumbs(style: :bootstrap)
+  end
+
+  test "custom container and fragment tags" do
+    breadcrumb :basic
+    assert_equal %{<c class="breadcrumbs"><f><a href="/">Home</a></f> &rsaquo; <f class="current">About</f></c>},
+                 breadcrumbs(container_tag: :c, fragment_tag: :f)
+  end
+
+  test "custom semantic container and fragment tags" do
+    breadcrumb :basic
+    assert_equal %{<c class="breadcrumbs"><f itemscope="" itemtype="http://data-vocabulary.org/Breadcrumb"><a href="/" itemprop="url"><span itemprop="title">Home</span></a></f> &rsaquo; <f class="current" itemscope="" itemtype="http://data-vocabulary.org/Breadcrumb"><span itemprop="title">About</span></f></c>},
+                 breadcrumbs(container_tag: :c, fragment_tag: :f, semantic: true)
+  end
+
+  test "unknown style" do
+    breadcrumb :basic
+    assert_raises ArgumentError do
+      breadcrumbs(style: :nonexistent)
+    end
+  end
+
+  test "register style" do
+    Gretel.register_style :test_style, { container_tag: :one, fragment_tag: :two }
+    
+    breadcrumb :basic
+
+    assert_equal %{<one class="breadcrumbs"><two><a href="/">Home</a></two><two class="current">About</two></one>},
+                 breadcrumbs(style: :test_style)
+  end
+
+  # Configuration reload
 
   test "reload configuration when file is changed" do
     path = setup_loading_from_tmp_folder
@@ -193,7 +337,7 @@ class HelperMethodsTest < ActionView::TestCase
     end
 
     breadcrumb :about
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &gt; <span class="current">About (loaded)</span></div>}, breadcrumbs
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &rsaquo; <span class="current">About (loaded)</span></div>}, breadcrumbs
 
     sleep 1 # File change interval is 1 second
 
@@ -209,7 +353,7 @@ class HelperMethodsTest < ActionView::TestCase
     end
 
     breadcrumb :about
-    assert_equal %{<div class="breadcrumbs"><a href="/test">Home (reloaded)</a> &gt; <span class="current">About (reloaded)</span></div>}, breadcrumbs
+    assert_equal %{<div class="breadcrumbs"><a href="/test">Home (reloaded)</a> &rsaquo; <span class="current">About (reloaded)</span></div>}, breadcrumbs
   end
 
   test "reload configuration when file is added" do
@@ -238,7 +382,7 @@ class HelperMethodsTest < ActionView::TestCase
     end
 
     breadcrumb :about
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &gt; <span class="current">About (loaded)</span></div>}, breadcrumbs
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &rsaquo; <span class="current">About (loaded)</span></div>}, breadcrumbs
   end
 
   test "reload configuration when file is deleted" do
@@ -266,7 +410,7 @@ class HelperMethodsTest < ActionView::TestCase
     end
 
     breadcrumb :contact
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &gt; <a href="/about">About (loaded)</a> &gt; <span class="current">Contact (loaded)</span></div>}, breadcrumbs
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &rsaquo; <a href="/about">About (loaded)</a> &rsaquo; <span class="current">Contact (loaded)</span></div>}, breadcrumbs
 
     File.delete path.join("pages.rb")
 
@@ -276,7 +420,7 @@ class HelperMethodsTest < ActionView::TestCase
     end
 
     breadcrumb :about
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &gt; <span class="current">About (loaded)</span></div>}, breadcrumbs
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &rsaquo; <span class="current">About (loaded)</span></div>}, breadcrumbs
   end
 
   test "reloads only in development environment" do
@@ -296,7 +440,7 @@ class HelperMethodsTest < ActionView::TestCase
     end
 
     breadcrumb :about
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &gt; <span class="current">About (loaded)</span></div>}, breadcrumbs
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &rsaquo; <span class="current">About (loaded)</span></div>}, breadcrumbs
 
     sleep 1
 
@@ -312,7 +456,7 @@ class HelperMethodsTest < ActionView::TestCase
     end
 
     breadcrumb :about
-    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &gt; <span class="current">About (loaded)</span></div>}, breadcrumbs
+    assert_equal %{<div class="breadcrumbs"><a href="/">Home (loaded)</a> &rsaquo; <span class="current">About (loaded)</span></div>}, breadcrumbs
   end
 
 private